[dev #601] move class Idea from Activity to root Folder
