+++ src/frontend/boilerplates/archipelago/src/resources/Idea/IdeaFilterSidebar.js
