+++ src/frontend/boilerplates/archipelago/src/resources/Idea/IdeaList.js
