+++ src/frontend/boilerplates/archipelago/src/resources/Idea/IdeaShow.js
