+++ src/frontend/boilerplates/archipelago/src/resources/Idea/IdeaTitle.js
